Refactor the code by transferring certain elements from BoringViewModel to BoringView Coordinator. This will help isolate the singleton-required code from the ViewModel that is now specific to each window.
